add a custom variable overwriting
This has to be done due to some issues during a project deployment.
Apparently, the variables define in the project's variable files are not
being picked up, in which ansible uses the default variable defined in
the default variables file. This branch will need to be removed once the
issue has been taken care of.